Fixed Destination if statement
Before this change, the code would only check if the Y axis was at the stated destination. It only fixed the X axis, but didn't make check if in the end it made it to the goal. So I changed the if statement to check for the X axis and the Y axis before completing the autonomous program.